refactor: move FileInfo/FileInfos definitions to daft-recordbatch
