+++ src/daft-logical-plan/src/lib.rs
@@ -30,7 +30,7 @@ use pyo3::prelude::*;
 #[cfg(feature = "python")]
 pub use sink_info::{CatalogType, DeltaLakeCatalogInfo, IcebergCatalogInfo, LanceCatalogInfo};
 pub use sink_info::{OutputFileInfo, SinkInfo};
-pub use source_info::{FileInfo, FileInfos, InMemoryInfo, SourceInfo};
+pub use source_info::{InMemoryInfo, SourceInfo};
 
 #[cfg(feature = "python")]
 pub fn register_modules(parent: &Bound<PyModule>) -> PyResult<()> {
@@ -41,8 +41,6 @@ pub fn register_modules(parent: &Bound<PyModule>) -> PyResult<()> {
     parent.add_class::<WarcSourceConfig>()?;
     parent.add_class::<CsvSourceConfig>()?;
     parent.add_class::<DatabaseSourceConfig>()?;
-    parent.add_class::<FileInfos>()?;
-    parent.add_class::<FileInfo>()?;
     parent.add_class::<JoinOptions>()?;
 
     Ok(())
