+++ src/daft-recordbatch/Cargo.toml
@@ -4,11 +4,11 @@ comfy-table = {workspace = true}
 common-arrow-ffi = {path = "../common/arrow-ffi", default-features = false}
 common-display = {path = "../common/display", default-features = false}
 common-error = {path = "../common/error", default-features = false}
+common-py-serde = {path = "../common/py-serde", default-features = false}
 common-runtime = {path = "../common/runtime", default-features = false}
 daft-core = {path = "../daft-core", default-features = false}
 daft-dsl = {path = "../daft-dsl", default-features = false}
 daft-image = {path = "../daft-image", default-features = false}
-daft-logical-plan = {path = "../daft-logical-plan", default-features = false}
 futures = {workspace = true}
 html-escape = {workspace = true}
 indexmap = {workspace = true}
@@ -19,7 +19,16 @@ serde = {workspace = true}
 serde_json = {workspace = true}
 
 [features]
-python = ["dep:pyo3", "common-error/python", "daft-core/python", "daft-dsl/python", "common-arrow-ffi/python", "common-display/python", "daft-image/python", "daft-logical-plan/python"]
+python = [
+  "dep:pyo3",
+  "common-error/python",
+  "daft-core/python",
+  "daft-dsl/python",
+  "common-arrow-ffi/python",
+  "common-display/python",
+  "common-py-serde/python",
+  "daft-image/python"
+]
 
 [lints]
 workspace = true
